test(governance): 外提运行时测试模块
