+++ src-tauri/src/codex/mod.rs
@@ -2051,479 +2051,5 @@ Task:\n{cleaned_prompt}"
 }
 
 #[cfg(test)]
-mod tests {
-    use super::thread_listing::{
-        build_local_codex_session_preview, build_thread_list_empty_response,
-        codex_session_identifier_candidates, merge_unified_codex_thread_entries,
-    };
-    use super::{create_session_runtime_recovering_error, run_start_thread_with_retry};
-    use crate::types::{LocalUsageSessionSummary, LocalUsageUsageData};
-    use serde_json::json;
-    use std::collections::HashSet;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::Arc;
-
-    #[test]
-    fn build_thread_list_empty_response_has_expected_shape() {
-        let response = build_thread_list_empty_response();
-        assert_eq!(response["result"]["data"], json!([]));
-        assert!(response["result"]["nextCursor"].is_null());
-    }
-
-    #[test]
-    fn build_local_codex_session_preview_prefers_trimmed_summary() {
-        let with_summary = build_local_codex_session_preview(
-            Some("  fixed preview  ".to_string()),
-            "openai/gpt-5".to_string(),
-        );
-        let without_summary =
-            build_local_codex_session_preview(Some("   ".to_string()), "openai/gpt-5".to_string());
-        assert_eq!(with_summary, "fixed preview");
-        assert_eq!(without_summary, "Codex session (openai/gpt-5)");
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_dedupes_and_keeps_metadata_stable() {
-        let live_entries = vec![
-            json!({
-                "id": "thread-live",
-                "preview": "live",
-                "updatedAt": 100,
-                "createdAt": 100
-            }),
-            json!({
-                "id": "thread-dup",
-                "preview": "remote",
-                "updatedAt": 90,
-                "createdAt": 90
-            }),
-            json!({
-                "id": "thread-dup",
-                "preview": "stale",
-                "updatedAt": 80,
-                "createdAt": 80
-            }),
-        ];
-        let local_sessions = vec![
-            LocalUsageSessionSummary {
-                session_id: "thread-dup".to_string(),
-                session_id_aliases: Vec::new(),
-                timestamp: 110,
-                cwd: None,
-                model: "openai/gpt-5".to_string(),
-                usage: LocalUsageUsageData::default(),
-                cost: 0.0,
-                summary: Some("local".to_string()),
-                source: Some("custom".to_string()),
-                provider: Some("openai".to_string()),
-                file_size_bytes: Some(4_096),
-                modified_lines: 0,
-            },
-            LocalUsageSessionSummary {
-                session_id: "thread-local".to_string(),
-                session_id_aliases: Vec::new(),
-                timestamp: 105,
-                cwd: None,
-                model: "openai/gpt-5-mini".to_string(),
-                usage: LocalUsageUsageData::default(),
-                cost: 0.0,
-                summary: Some("local-only".to_string()),
-                source: Some("project".to_string()),
-                provider: Some("openai".to_string()),
-                file_size_bytes: Some(8_192),
-                modified_lines: 0,
-            },
-        ];
-
-        let workspace_session_ids: HashSet<String> = local_sessions
-            .iter()
-            .flat_map(codex_session_identifier_candidates)
-            .collect();
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &local_sessions,
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-
-        assert_eq!(merged.len(), 3);
-        assert_eq!(merged[0]["id"], "thread-dup");
-        assert_eq!(merged[0]["updatedAt"], 110);
-        assert_eq!(merged[0]["preview"], "remote");
-        assert_eq!(merged[0]["sizeBytes"], 4_096);
-        assert_eq!(merged[0]["source"], "custom");
-        assert_eq!(merged[0]["provider"], "openai");
-        assert_eq!(merged[0]["sourceLabel"], "custom/openai");
-
-        assert_eq!(merged[1]["id"], "thread-local");
-        assert_eq!(merged[1]["localFallback"], true);
-        assert_eq!(merged[1]["sizeBytes"], 8_192);
-        assert_eq!(merged[1]["sourceLabel"], "project/openai");
-
-        assert_eq!(merged[2]["id"], "thread-live");
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_replaces_generic_vscode_source() {
-        let live_entries = vec![json!({
-            "id": "thread-dup",
-            "preview": "remote",
-            "updatedAt": 90,
-            "createdAt": 90,
-            "source": "vscode",
-            "sourceLabel": "vscode"
-        })];
-        let local_sessions = vec![LocalUsageSessionSummary {
-            session_id: "thread-dup".to_string(),
-            session_id_aliases: Vec::new(),
-            timestamp: 110,
-            cwd: None,
-            model: "openai/gpt-5".to_string(),
-            usage: LocalUsageUsageData::default(),
-            cost: 0.0,
-            summary: Some("local".to_string()),
-            source: Some("mossx".to_string()),
-            provider: Some("openai".to_string()),
-            file_size_bytes: Some(1_024),
-            modified_lines: 0,
-        }];
-
-        let workspace_session_ids: HashSet<String> = local_sessions
-            .iter()
-            .flat_map(codex_session_identifier_candidates)
-            .collect();
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &local_sessions,
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0]["sizeBytes"], 1_024);
-        assert_eq!(merged[0]["source"], "mossx");
-        assert_eq!(merged[0]["sourceLabel"], "mossx/openai");
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_matches_session_id_aliases() {
-        let live_entries = vec![json!({
-            "id": "rollout-2026-04-10T10-00-00-session-123",
-            "preview": "remote",
-            "updatedAt": 90,
-            "createdAt": 90
-        })];
-        let local_sessions = vec![LocalUsageSessionSummary {
-            session_id: "session-123".to_string(),
-            session_id_aliases: vec!["rollout-2026-04-10T10-00-00-session-123".to_string()],
-            timestamp: 110,
-            cwd: None,
-            model: "openai/gpt-5".to_string(),
-            usage: LocalUsageUsageData::default(),
-            cost: 0.0,
-            summary: Some("local".to_string()),
-            source: Some("cli".to_string()),
-            provider: Some("openai".to_string()),
-            file_size_bytes: Some(2_048),
-            modified_lines: 0,
-        }];
-
-        let workspace_session_ids: HashSet<String> = local_sessions
-            .iter()
-            .flat_map(codex_session_identifier_candidates)
-            .collect();
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &local_sessions,
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0]["id"], "rollout-2026-04-10T10-00-00-session-123");
-        assert_eq!(merged[0]["sizeBytes"], 2_048);
-        assert_eq!(merged[0]["source"], "cli");
-        assert_eq!(merged[0]["sourceLabel"], "cli/openai");
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_filters_background_helper_sessions() {
-        let live_entries = vec![
-            json!({
-                "id": "thread-memory-helper",
-                "preview": "live row should be hidden through local alias",
-                "updatedAt": 120,
-                "createdAt": 120
-            }),
-            json!({
-                "id": "thread-title-helper",
-                "preview": "Generate a concise title for a coding chat thread from the first user message. Return only title text.",
-                "updatedAt": 115,
-                "createdAt": 115
-            }),
-            json!({
-                "id": "thread-visible",
-                "preview": "normal user prompt",
-                "updatedAt": 100,
-                "createdAt": 100
-            }),
-        ];
-        let local_sessions = vec![
-            LocalUsageSessionSummary {
-                session_id: "session-memory-helper".to_string(),
-                session_id_aliases: vec!["thread-memory-helper".to_string()],
-                timestamp: 125,
-                cwd: None,
-                model: "openai/gpt-5".to_string(),
-                usage: LocalUsageUsageData::default(),
-                cost: 0.0,
-                summary: Some(
-                    "## Memory Writing Agent: Phase 2 (Consolidation)\n\nConsolidate raw memories."
-                        .to_string(),
-                ),
-                source: Some("cli".to_string()),
-                provider: Some("openai".to_string()),
-                file_size_bytes: Some(2_048),
-                modified_lines: 0,
-            },
-            LocalUsageSessionSummary {
-                session_id: "thread-visible-local".to_string(),
-                session_id_aliases: Vec::new(),
-                timestamp: 90,
-                cwd: None,
-                model: "openai/gpt-5".to_string(),
-                usage: LocalUsageUsageData::default(),
-                cost: 0.0,
-                summary: Some("normal local prompt".to_string()),
-                source: Some("cli".to_string()),
-                provider: Some("openai".to_string()),
-                file_size_bytes: Some(1_024),
-                modified_lines: 0,
-            },
-        ];
-
-        let workspace_session_ids: HashSet<String> = local_sessions
-            .iter()
-            .flat_map(codex_session_identifier_candidates)
-            .collect();
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &local_sessions,
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-        let ids = merged
-            .iter()
-            .filter_map(|entry| entry.get("id").and_then(|value| value.as_str()))
-            .collect::<Vec<_>>();
-
-        assert_eq!(ids, vec!["thread-visible", "thread-visible-local"]);
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_does_not_backfill_cwd_for_unmapped_live_rows() {
-        let live_entries = vec![json!({
-            "id": "thread-live",
-            "preview": "remote",
-            "updatedAt": 90,
-            "createdAt": 90
-        })];
-
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &[],
-            &HashSet::new(),
-            "/tmp/workspace",
-            10,
-        );
-
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0]["id"], "thread-live");
-        assert!(merged[0].get("cwd").is_none() || merged[0]["cwd"].is_null());
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_backfills_cwd_from_cached_workspace_ids() {
-        let live_entries = vec![json!({
-            "id": "thread-live",
-            "preview": "remote",
-            "updatedAt": 90,
-            "createdAt": 90
-        })];
-        let mut workspace_session_ids = HashSet::new();
-        workspace_session_ids.insert("thread-live".to_string());
-
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &[],
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0]["id"], "thread-live");
-        assert_eq!(merged[0]["cwd"], "/tmp/workspace");
-    }
-
-    #[test]
-    fn merge_unified_codex_thread_entries_backfills_workspace_cwd_for_mapped_live_rows() {
-        let live_entries = vec![json!({
-            "id": "rollout-2026-04-10T10-00-00-session-123",
-            "preview": "remote",
-            "updatedAt": 90,
-            "createdAt": 90
-        })];
-        let local_sessions = vec![LocalUsageSessionSummary {
-            session_id: "session-123".to_string(),
-            session_id_aliases: vec!["rollout-2026-04-10T10-00-00-session-123".to_string()],
-            timestamp: 110,
-            cwd: None,
-            model: "openai/gpt-5".to_string(),
-            usage: LocalUsageUsageData::default(),
-            cost: 0.0,
-            summary: Some("local".to_string()),
-            source: Some("cli".to_string()),
-            provider: Some("openai".to_string()),
-            file_size_bytes: Some(2_048),
-            modified_lines: 0,
-        }];
-
-        let workspace_session_ids: HashSet<String> = local_sessions
-            .iter()
-            .flat_map(codex_session_identifier_candidates)
-            .collect();
-        let merged = merge_unified_codex_thread_entries(
-            live_entries,
-            &local_sessions,
-            &workspace_session_ids,
-            "/tmp/workspace",
-            10,
-        );
-
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0]["id"], "rollout-2026-04-10T10-00-00-session-123");
-        assert_eq!(merged[0]["cwd"], "/tmp/workspace");
-    }
-
-    #[tokio::test]
-    async fn start_thread_retry_reacquires_after_manual_shutdown_race() {
-        let ensure_calls = Arc::new(AtomicUsize::new(0));
-        let start_calls = Arc::new(AtomicUsize::new(0));
-
-        let result = run_start_thread_with_retry(
-            "ws-1",
-            {
-                let ensure_calls = Arc::clone(&ensure_calls);
-                move || {
-                    let ensure_calls = Arc::clone(&ensure_calls);
-                    async move {
-                        ensure_calls.fetch_add(1, Ordering::SeqCst);
-                        Ok(())
-                    }
-                }
-            },
-            {
-                let start_calls = Arc::clone(&start_calls);
-                move || {
-                    let start_calls = Arc::clone(&start_calls);
-                    async move {
-                        let attempt = start_calls.fetch_add(1, Ordering::SeqCst);
-                        if attempt == 0 {
-                            Err(
-                                "[RUNTIME_ENDED] Managed runtime stopped after manual shutdown."
-                                    .to_string(),
-                            )
-                        } else {
-                            Ok(json!({ "result": { "threadId": "thread-recovered" } }))
-                        }
-                    }
-                }
-            },
-        )
-        .await
-        .expect("manual shutdown race should retry once");
-
-        assert_eq!(result["result"]["threadId"], "thread-recovered");
-        assert_eq!(ensure_calls.load(Ordering::SeqCst), 2);
-        assert_eq!(start_calls.load(Ordering::SeqCst), 2);
-    }
-
-    #[tokio::test]
-    async fn start_thread_retry_does_not_retry_non_runtime_shutdown_errors() {
-        let ensure_calls = Arc::new(AtomicUsize::new(0));
-        let start_calls = Arc::new(AtomicUsize::new(0));
-
-        let error = run_start_thread_with_retry(
-            "ws-1",
-            {
-                let ensure_calls = Arc::clone(&ensure_calls);
-                move || {
-                    let ensure_calls = Arc::clone(&ensure_calls);
-                    async move {
-                        ensure_calls.fetch_add(1, Ordering::SeqCst);
-                        Ok(())
-                    }
-                }
-            },
-            {
-                let start_calls = Arc::clone(&start_calls);
-                move || {
-                    let start_calls = Arc::clone(&start_calls);
-                    async move {
-                        start_calls.fetch_add(1, Ordering::SeqCst);
-                        Err("workspace not connected".to_string())
-                    }
-                }
-            },
-        )
-        .await
-        .expect_err("non-runtime errors should surface directly");
-
-        assert_eq!(error, "workspace not connected");
-        assert_eq!(ensure_calls.load(Ordering::SeqCst), 1);
-        assert_eq!(start_calls.load(Ordering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn start_thread_retry_returns_recoverable_error_when_stopping_race_persists() {
-        let ensure_calls = Arc::new(AtomicUsize::new(0));
-        let start_calls = Arc::new(AtomicUsize::new(0));
-
-        let error = run_start_thread_with_retry(
-            "ws-1",
-            {
-                let ensure_calls = Arc::clone(&ensure_calls);
-                move || {
-                    let ensure_calls = Arc::clone(&ensure_calls);
-                    async move {
-                        ensure_calls.fetch_add(1, Ordering::SeqCst);
-                        Ok(())
-                    }
-                }
-            },
-            {
-                let start_calls = Arc::clone(&start_calls);
-                move || {
-                    let start_calls = Arc::clone(&start_calls);
-                    async move {
-                        start_calls.fetch_add(1, Ordering::SeqCst);
-                        Err(
-                            "[RUNTIME_ENDED] Managed runtime stopped after manual shutdown."
-                                .to_string(),
-                        )
-                    }
-                }
-            },
-        )
-        .await
-        .expect_err("persistent stopping race should surface recoverable error");
-
-        assert_eq!(error, create_session_runtime_recovering_error());
-        assert_eq!(ensure_calls.load(Ordering::SeqCst), 2);
-        assert_eq!(start_calls.load(Ordering::SeqCst), 2);
-    }
-}
+#[path = "codex_tests.rs"]
+mod tests;
